+++ android/src/main/kotlin/com/github/florent37/assets_audio_player/Player.kt
@@ -8,15 +8,13 @@ import android.os.Handler
 import android.os.Message
 import com.github.florent37.assets_audio_player.notification.AudioMetas
 import com.github.florent37.assets_audio_player.notification.NotificationManager
+import com.github.florent37.assets_audio_player.notification.NotificationSettings
 import com.google.android.exoplayer2.*
 import com.google.android.exoplayer2.Player
 import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory
 import com.google.android.exoplayer2.source.MediaSource
 import com.google.android.exoplayer2.source.ProgressiveMediaSource
-import com.google.android.exoplayer2.upstream.AssetDataSource
-import com.google.android.exoplayer2.upstream.DataSource
-import com.google.android.exoplayer2.upstream.DataSpec
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory
+import com.google.android.exoplayer2.upstream.*
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.MethodChannel
 import kotlin.math.max
@@ -77,6 +75,7 @@ class Player(
 
     private var displayNotification = false
     private var audioMetas: AudioMetas? = null
+    private var notificationSettings: NotificationSettings? = null
 
     private val updatePosition = object : Runnable {
         override fun run() {
@@ -124,6 +123,7 @@ class Player(
              seek: Int?,
              respectSilentMode: Boolean,
              displayNotification: Boolean,
+             notificationSettings: NotificationSettings,
              audioMetas: AudioMetas,
              playSpeed: Double,
              result: MethodChannel.Result,
@@ -134,6 +134,7 @@ class Player(
         this.mediaPlayer = SimpleExoPlayer.Builder(context).build();
         this.displayNotification = displayNotification
         this.audioMetas = audioMetas
+        this.notificationSettings = notificationSettings
         this.respectSilentMode = respectSilentMode
 
         lateinit var mediaSource: MediaSource
@@ -145,7 +146,7 @@ class Player(
                         .createMediaSource(Uri.parse(assetAudioPath))
             } else if (audioType == AUDIO_TYPE_FILE) {
                 mediaSource = ProgressiveMediaSource
-                        .Factory(DefaultDataSourceFactory(context, "assets_audio_player"), DefaultExtractorsFactory())
+                        .Factory(FileDataSource.Factory(), DefaultExtractorsFactory())
                         .createMediaSource(Uri.parse(assetAudioPath))
             } else { //asset
                 val path = if (assetAudioPackage.isNullOrBlank()) {
@@ -269,7 +270,14 @@ class Player(
 
     private fun updateNotif() {
         this.audioMetas?.takeIf { this.displayNotification }?.let { audioMetas ->
-            notificationManager.showNotification(playerId = id, audioMetas = audioMetas, isPlaying = this.isPlaying)
+            this.notificationSettings?.let { notificationSettings ->
+                notificationManager.showNotification(
+                        playerId = id,
+                        audioMetas = audioMetas,
+                        isPlaying = this.isPlaying,
+                        notificationSettings = notificationSettings
+                )
+            }
         }
     }