+++ android/src/main/kotlin/com/github/florent37/assets_audio_player/notification/NotificationManager.kt
@@ -6,19 +6,25 @@ import com.github.florent37.assets_audio_player.AssetsAudioPlayerPlugin
 
 class NotificationManager(private val context: Context) {
 
-    fun showNotification(playerId: String, audioMetas: AudioMetas, isPlaying: Boolean) {
+    var closed = false
+
+    fun showNotification(playerId: String, audioMetas: AudioMetas, isPlaying: Boolean, notificationSettings: NotificationSettings) {
+        if(closed)
+            return
         context.startService(Intent(context, NotificationService::class.java).apply {
             putExtra(NotificationService.EXTRA_NOTIFICATION_ACTION, NotificationAction.Show(
                     isPlaying = isPlaying,
                     audioMetas = audioMetas,
-                    playerId = playerId
+                    playerId = playerId,
+                    notificationSettings= notificationSettings
             ))
         })
         AssetsAudioPlayerPlugin.instance?.assetsAudioPlayer?.registerLastPlayerWithNotif(playerId)
     }
 
-    fun hideNotification() {
+    fun hideNotification(definitively: Boolean = false) {
         //if remainingNotif == 0, stop
         context.stopService(Intent(context, NotificationService::class.java))
+        closed = definitively
     }
 }