+++ lib/src/assets_audio_player.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:math';
 
+import 'package:assets_audio_player/src/notification.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:rxdart/rxdart.dart';
@@ -9,9 +10,11 @@ import 'package:uuid/uuid.dart';
 
 import 'playable.dart';
 import 'playing.dart';
+import 'notification.dart';
 
 export 'playable.dart';
 export 'playing.dart';
+export 'notification.dart';
 
 const _DEFAULT_AUTO_START = true;
 const _DEFAULT_RESPECT_SILENT_MODE = false;
@@ -25,9 +28,10 @@ const METHOD_IS_PLAYING = "player.isPlaying";
 const METHOD_IS_BUFFERING = "player.isBuffering";
 const METHOD_CURRENT = "player.current";
 const METHOD_FORWARD_REWIND_SPEED = "player.forwardRewind";
-const METHOD_NEXT = "player.next";
-const METHOD_PREV = "player.prev";
-const METHOD_PLAY_OR_PAUSE = "player.playOrPause";
+const METHOD_NOTIFICATION_NEXT = "player.next";
+const METHOD_NOTIFICATION_PREV = "player.prev";
+const METHOD_NOTIFICATION_STOP = "player.stop";
+const METHOD_NOTIFICATION_PLAY_OR_PAUSE = "player.playOrPause";
 const METHOD_PLAY_SPEED = "player.playSpeed";
 
 /// The AssetsAudioPlayer, playing audios from assets/
@@ -51,6 +55,8 @@ class AssetsAudioPlayer {
   static final double maxPlaySpeed = 16.0;
   static final double defaultVolume = maxVolume;
   static final double defaultPlaySpeed = 1.0;
+  static final NotificationSettings defaultNotificationSettings =
+      const NotificationSettings();
 
   static final uuid = Uuid();
 
@@ -317,14 +323,17 @@ class AssetsAudioPlayer {
         case METHOD_FINISHED:
           _onFinished(call.arguments);
           break;
-        case METHOD_NEXT:
-          next();
+        case METHOD_NOTIFICATION_NEXT:
+          _notificationNext();
           break;
-        case METHOD_PREV:
-          previous();
+        case METHOD_NOTIFICATION_PREV:
+          _notificationPrevious();
           break;
-        case METHOD_PLAY_OR_PAUSE: //eg: from notification
-          playOrPause();
+        case METHOD_NOTIFICATION_STOP:
+          _notificationStop();
+          break;
+        case METHOD_NOTIFICATION_PLAY_OR_PAUSE: //eg: from notification
+          _notificationPlayPause();
           break;
         case METHOD_CURRENT:
           if (call.arguments == null) {
@@ -431,7 +440,8 @@ class AssetsAudioPlayer {
     return false;
   }
 
-  Future<void> _openPlaylistCurrent() async {
+  Future<void> _openPlaylistCurrent(
+      {bool autoStart = true, Duration seek}) async {
     if (_playlist != null) {
       return _open(
         _playlist.currentAudio(),
@@ -439,6 +449,9 @@ class AssetsAudioPlayer {
         respectSilentMode: _playlist.respectSilentMode,
         showNotification: _playlist.showNotification,
         playSpeed: _playlist.playSpeed,
+        notificationSettings: _playlist.notificationSettings,
+        autoStart: autoStart,
+        seek: seek,
       );
     }
   }
@@ -528,6 +541,34 @@ class AssetsAudioPlayer {
     }
   }
 
+  void _notificationPrevious() {
+    if (_playlist?.notificationSettings?.customPrevAction != null) {
+      _playlist?.notificationSettings?.customPrevAction(this);
+    } else {
+      previous();
+    }
+  }
+
+  void _notificationStop() {
+    stop();
+  }
+
+  void _notificationPlayPause() {
+    if (_playlist?.notificationSettings?.customPlayPauseAction != null) {
+      _playlist?.notificationSettings?.customPlayPauseAction(this);
+    } else {
+      playOrPause();
+    }
+  }
+
+  void _notificationNext() {
+    if (_playlist?.notificationSettings?.customNextAction != null) {
+      _playlist?.notificationSettings?.customNextAction(this);
+    } else {
+      next();
+    }
+  }
+
   //private method, used in open(playlist) and open(path)
   Future<void> _open(
     Audio audio, {
@@ -537,6 +578,7 @@ class AssetsAudioPlayer {
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
     Duration seek,
     double playSpeed,
+    NotificationSettings notificationSettings,
   }) async {
     final currentAudio = _lastOpenedAssetsAudio;
     if (audio != null) {
@@ -558,19 +600,14 @@ class AssetsAudioPlayer {
         if (audio.package != null) {
           params["package"] = audio.package;
         }
+
+        //region notifs
+        final notifSettings = notificationSettings ?? NotificationSettings();
+        writeNotificationSettingsInto(params, notifSettings);
+        //endregion
+
         if (audio.metas != null) {
-          if (audio.metas.title != null)
-            params["song.title"] = audio.metas.title;
-          if (audio.metas.artist != null)
-            params["song.artist"] = audio.metas.artist;
-          if (audio.metas.album != null)
-            params["song.album"] = audio.metas.album;
-          if (audio.metas.image != null) {
-            params["song.image"] = audio.metas.image.path;
-            params["song.imageType"] = audio.metas.image.type.description();
-            if (audio.metas.image.package != null)
-              params["song.imagePackage"] = audio.metas.image.package;
-          }
+          writeAudioMetasInto(params, audio.metas);
         }
         _lastOpenedAssetsAudio = audio;
         /*final result = */
@@ -594,27 +631,20 @@ class AssetsAudioPlayer {
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
     Duration seek,
     double playSpeed,
+    NotificationSettings notificationSettings,
   }) async {
     _lastSeek = null;
     _replaceRealtimeSubscription();
     this._playlist = _CurrentPlaylist(
-      playlist: playlist,
-      volume: volume,
-      respectSilentMode: respectSilentMode,
-      showNotification: showNotification,
-      playSpeed: playSpeed,
-    );
+        playlist: playlist,
+        volume: volume,
+        respectSilentMode: respectSilentMode,
+        showNotification: showNotification,
+        playSpeed: playSpeed,
+        notificationSettings: notificationSettings);
     _playlist.clearPlayeAudio(shuffle);
     _playlist.moveTo(playlist.startIndex);
-    return _open(
-      _playlist.currentAudio(),
-      autoStart: autoStart,
-      forcedVolume: volume,
-      playSpeed: playSpeed,
-      respectSilentMode: respectSilentMode,
-      showNotification: showNotification,
-      seek: seek,
-    );
+    return _openPlaylistCurrent(autoStart: autoStart, seek: seek);
   }
 
   bool get _isLiveStream {
@@ -642,29 +672,30 @@ class AssetsAudioPlayer {
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
     Duration seek,
     double playSpeed,
+    NotificationSettings notificationSettings,
   }) async {
     if (playable is Playlist &&
         playable.audios != null &&
         playable.audios.length > 0) {
-      await _openPlaylist(
-        playable,
-        autoStart: autoStart,
-        volume: volume,
-        respectSilentMode: respectSilentMode,
-        showNotification: showNotification,
-        seek: seek,
-        playSpeed: playSpeed,
-      );
+      await _openPlaylist(playable,
+          autoStart: autoStart,
+          volume: volume,
+          respectSilentMode: respectSilentMode,
+          showNotification: showNotification,
+          seek: seek,
+          playSpeed: playSpeed,
+          notificationSettings:
+              notificationSettings ?? defaultNotificationSettings);
     } else if (playable is Audio) {
-      await _openPlaylist(
-        Playlist(audios: [playable]),
-        autoStart: autoStart,
-        volume: volume,
-        respectSilentMode: respectSilentMode,
-        showNotification: showNotification,
-        seek: seek,
-        playSpeed: playSpeed,
-      );
+      await _openPlaylist(Playlist(audios: [playable]),
+          autoStart: autoStart,
+          volume: volume,
+          respectSilentMode: respectSilentMode,
+          showNotification: showNotification,
+          seek: seek,
+          playSpeed: playSpeed,
+          notificationSettings:
+              notificationSettings ?? defaultNotificationSettings);
     } else {
       //do nothing
       //throw exception ?
@@ -858,6 +889,7 @@ class _CurrentPlaylist {
   final bool respectSilentMode;
   final bool showNotification;
   final double playSpeed;
+  final NotificationSettings notificationSettings;
 
   int playlistIndex = 0;
 
@@ -953,6 +985,7 @@ class _CurrentPlaylist {
     this.respectSilentMode,
     this.showNotification,
     this.playSpeed,
+    this.notificationSettings,
   });
 
   void returnToFirst() {